Search a sequence of arrays of FuseProps without using a key path
